Move startSupervision to BridgeDetailFooter
